backend/requirements.txt
asgiref==3.7.2
cog==0.7.2
colorama==0.4.6
cryptography==40.0.2
decord==0.6.0
fairscale==0.4.4
numpy==1.24.3
openai==0.27.6
opencv_python_headless==4.5.5.64
Pillow==9.5.0
Pillow==9.5.0
pycocoevalcap==1.2
pycocotools==2.0.6
python-dotenv==1.0.0
replicate==0.8.3
Requests==2.31.0
requests_html==0.10.0
ruamel_yaml==0.17.29
setuptools==52.0.0
timm==0.4.12
torch==2.0.0+cpu
torchvision==0.15.1+cpu
transformers==4.28.1
typing_extensions==4.6.2
watchdog==3.0.0
